Refactor: Move component classes only used by ChatMessageContainer classes into same package
