+++ apps/desktop/desktop/src/main/java/bisq/desktop/main/content/chat/message_container/components/CitationBlock.java
@@ -15,7 +15,7 @@
  * along with Bisq. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.desktop.main.content.components;
+package bisq.desktop.main.content.chat.message_container.components;
 
 import bisq.chat.ChatMessage;
 import bisq.chat.ChatService;
